Added stop watch to logging. Also corrected namespaces of sorted after they were moved in last update.
